faker seed

+++ main/infrastructure/controllers/mapper/json_schema.go
@@ -15,7 +15,7 @@ type JsonProperty struct {
 	Type       properties.Type `json:"type"`
 	Min        int             `json:"min,omitempty"`
 	Max        int             `json:"max,omitempty"`
-	Rate       float32         `json:"rate,omitempty"`
+	Rate       int             `json:"rate,omitempty"`
 	From       time.Time       `json:"from,omitempty"`
 	To         time.Time       `json:"to,omitempty"`
 	Element    *JsonProperty   `json:"element,omitempty"`
